pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ewankb"
version = "0.1.11"
description = "Build structured knowledge bases from documents and code, with graph-based querying."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
authors = [
    {name = "Ewan", email = "475299613@qq.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "graphifyy>=0.4.0",
    "anthropic>=0.40.0",
    "networkx>=3.0",
    "rank-bm25>=0.2.2",
    "jieba>=0.42",
    'tomli>=2.0; python_version < "3.11"',
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[project.scripts]
ewankb = "ewankb.__main__:main"

[project.urls]
Homepage = "https://github.com/Ewan-Jone/ewan-kb"
Documentation = "https://github.com/Ewan-Jone/ewan-kb#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["ewankb*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
ewankb = ["skills/**/*.md", "templates/**/*.md", "templates/**/*.json",
          "tools/*.json", "tools/discover/*.json", "tools/fetch_repos/*.json",
          "tools/scrape_cf/*.json", "tools/project_config.template.json"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
